--- tests/python/integration_tests/scf/__init__.py
@@ -1,32 +1,0 @@
-#
-# Copyright 2023 NWChemEx-Project
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-import pluginplay
-import chemist
-import simde
-import nwchemex
-import unittest
-
-
-class TestSCF(unittest.TestCase):
-
-    def test_scf_energy(self):
-
-        mm = pluginplay.ModuleManager()
-        nwchemex.load_modules(mm)
-        mod = mm.at('SCF Wavefunction')
-
-        # TODO: Test when property type is exposed
